Added force close ability
Project Tracker sometimes runs in the background and would need to be force closed before it can be uninstalled. The functionality is there if we need it, but at the moment I solved it by removing server connectivity which was keeping the program alive even though we're not using a server for usage statistics at the moment.